+++ NEWS.md
@@ -1,3 +1,5 @@
+* Added the function `todos`.
+
 # add2todo 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.